+++ R/as_bootstrap_design.R
@@ -50,7 +50,8 @@
 #' }
 #' @return
 #' A replicate design object, with class \code{svyrep.design}, which can be used with the usual functions,
-#' such as \code{svymean()} or \code{svyglm()}. \cr \cr
+#' such as \code{svymean()} or \code{svyglm()}.
+#'
 #' Use \code{weights(..., type = 'analysis')} to extract the matrix of replicate weights. \cr
 #' Use \code{as_data_frame_with_weights()} to convert the design object to a data frame with columns
 #' for the full-sample and replicate weights.
@@ -215,7 +216,14 @@ as_bootstrap_design.survey.design <- function(design,
     rep_design <- survey::as.svrepdesign(design = design,
                                          type = type,
                                          compress = compress,
-                                         mse = mse)
+                                         mse = mse,
+                                         replicates = replicates)
+  }
+
+  if (inherits(design, 'tbl_svy') && ('package:srvyr' %in% search())) {
+    rep_design <- srvyr::as_survey_rep(
+      rep_design
+    )
   }
 
   rep_design$call <- sys.call(which = -1)